requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.3
msal==1.26.0
gunicorn==21.2.0

